Start of indexing work.
Keep track of bits written and the positions of every nth entry.

First plan is to do the simplest thing that could work - just write tuples
(num, bitpos) to the end of the file, ending with a footer saying where the
start is.  Users can then just interpolation search the index.

I'm considering a fancier variant; GCS the index, since it's going to
have the same properties, it will compress well.  Then index *that*,
such that a given entry can be found with a reasonably small number of
decodes.

This of course means encoding tuples, so it would take more code changes
than I'd like.  I'm not sure it'll be worth it given even a million
index points is all of 16MB.